canmsg fix

+++ Src/vesc_can.c
@@ -167,7 +167,7 @@ void comm_can_transmit_eid(uint32_t id, const uint8_t *data, uint8_t len)
     }
 
 #if VESC_CAN_ENABLE
-    can_ext_send_msg((uint16_t) id,(can_msg*) data); //须改为自己的can发送函数，拓展帧
+    can_ext_send_msg( id,(can_msg*) data); //须改为自己的can发送函数，拓展帧
 #else
     (void)id;
     (void)data;
